remove chart specific methods, consolidate others
