+++ src/index.js
@@ -1,20 +1,14 @@
-import Chart from "./chart";
 import * as Collection from "./collection";
-import Data from "./data";
-import Domain from "./domain";
+import Helpers from "./helpers";
 import * as Log from "./log";
-import Scale from "./scale";
 import * as Style from "./style";
 import * as Type from "./type";
 import * as PropTypes from "./prop-types";
 
 export default {
-  Chart,
   Collection,
-  Data,
-  Domain,
+  Helpers,
   Log,
-  Scale,
   Style,
   Type,
   PropTypes
